Made reusable model
The profanity filtering can be added by extending ProfFiltered_ModelForm instead of ModelForm. Mixin behavior would be better, but I need first dibs on __init__, I didn't want to have to make the subclass implement __init__ any differently, and I can't figure out how to get 'curried' generics-based inheritance goign like there can be in C++. Once functionality is added, you just need to pass override_text for the value of the overriding submit button, name the original and override buttons submit_btn (and add the whole error bar thing maybe), and add clean methods eg clean_x(self): return self.profanity_cleaner('x').
